move about endpoint to about package
