+++ traffic_ops/traffic_ops_golang/about/about.go
@@ -1,4 +1,4 @@
-package utils
+package about
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
